Refactor SendDataButton into buttons for live and set modes
